Revert "Refactor hooks and ProjectStats into seperate files"
This reverts commit ce73c96624d6490a53ea2001f7f1b65024b66d87.
